+++ MatrixBenchmarkCs.Inc/MultiplyMatrix/MatrixNMultiplyBenchmark.cs
@@ -138,7 +138,7 @@ namespace MatrixBenchmarkCs.MultiplyMatrix {
         };
 
         /// <inheritdoc cref="ILoopCountGetter.LoopCount" />
-        public int LoopCount { get { return N * N; } set { } }
+        public long LoopCount { get { return N * N; } set { } }
 
         /// <summary>The number of columns in matrix A, or the number of rows in matrix B (矩阵A的列数, 或矩阵B的行数).</summary>
         public int MatrixK { get { return N; } }
